+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 ## [2.0.0]
 * Added missing data types
 * Added Models for each data type
-* Updated getRecords to return Records instead of Maps
+* Added getRecords to return a List of Records instead of a Map 
 * Added writeData method
 * Added deleteRecords methods
+* Upgrade Health Connect API to `1.1.0-alpha2`
+* Updated README.md
 ## [1.2.3]
 * Upgrade to `alpha11` [#8](https://github.com/duynguyen242/flutter_health_connect/pull/8) by [aljkor](https://github.com/aljkor)
 ## [1.2.2]